Rework game logic: introduce points, early answers, etc
Points: Instead of trying to minimize distance, a points score is
calculated from 0 to 1000 by an inverse exponential function, so that
the score decreases fast from 1000 but than slows down. This way it does
not hurt too bad if you are really off on one question.

Now that points are used, we can give missing scores a score of 0. A
missing answer is if the user drops out or simply forgets to click on
the map before the deadline.

Answers can now be delivered multiple times instead of once on deadline
ended. That was too flaky. If that one call failed, the quiz could not
proceed. Now we can also be stricter on the deadline.